Updated camera definitions for DECam
I added the code, input files, and resulting camera registry
for the DECam.  You can load the camera by doing:
>>> from lsst.obs.decam import DecamMapper
>>> mapper = DecamMapper()
>>> mapper.camera

chipcenters.txt has the bounding boxes of the chips and a rotation in
the plane of the focal plane.
segmentfile.txt has the amp information.  This is trivial until we
can look at DECam raw data.