Make ingestImagesDecam.py accept raw files
using command-line option --filetype.
The hdu extension is added into the registry for Butler
to retrieve the corresponging hdu given ccdnum.
There was a mixed name use of ccdnum and HDU extension
number for the MEF fits, and these two numbers are not
the same for raw data. Further, the ordering of ccdnum
in fits hdu can change over time for DECam.