adds registry-free butler support
regarding the steps in the readme.md for this package,
now processCcdDecam.py can be run without the registry.sqlite file.
HOWEVER for this script, the data must still be imported (step 3,
ingestImagesDecam.py) because this creates files (aliases, actually)
that conform to the template in the policy file. The registry free
butler depends on this correlation between file names and template.