Removed the obs_decam config override for calibrate.initialPsf.fwhm
obs_decam config override has calibrate.initialPsf.fwhm=2.0 arcsec.
This is far too large. We decided on the RFD discussion on "initial
PSF FWHM" to remove this config override and use the default value
in the calibrate task. The idea is to tweak the calibrate default
value to make it work pretty much for all data.