Move defect registry into the package.
This is basically a copy of obs_subaru's method of generating the defect
registry.

This currently has a bug where it calls "ingestCalibs.py --create", but on
subsequent scons runs it will complain that the registry already exists. I think
this can be avoided by some more careful definition of the build process in
SConscript.