Add LSST linearizers
Add the standard DECam linearity table (a FITS file
with one binary table HDU per detector).
Add a script to generate LSST linearizers (as pickles
in decam/linearizer) from the standard DECam linearity table.
Modify two scons files so the LSST linearizers are automatically
generated when the package is built (if needed), and that this occurs
before unit tests are run.
Add decam/README.md to explain what is going on.
Update the DECam camera mapper to support the "linearizer" dataset type.
Add a unit test that shows the linearizers can be retrieved
using the butler.