Use HDU in instcal templates instead of ccdnum.
Since an HDU's position in an instcal file does not necessarily correspond to
its CCD number, this template had been mislabling CCDs.

Also removes special handling of decam raw files that was required to track the
hdu field. This code was largely a duplication of the pipe_tasks ingest task,
and can be alleviated with the minor changes to that task (also implemented on
this ticket.)