testProcessCcd: only run ProcessCcdTask once
By putting the use of ProcessCcdTask in the setUp method, it gets
run once for each of the 'test' methods (currently 3) which is
unnecessary. Moving it into setUpClass allows us to run it only
once, and the test methods can share the results; and none of the
tests fire if the testdata isn't available.