modernize Butler API use in testGetId
This test was failing because of a sequence of
events:

The test was passing into Butler init an
instantiated mapper passed in but not comparing
equal with the mapper loaded by the repository
cfg.

A repositoryCfg.yaml was being created by the old
use of the Butler API (because the old API creates
a read-write output repository), and it was not
geting cleaned up by the test.

When the test runs again on the same computer
where the old repositoryCfg.yaml file exists, the
mapper is reinstantiated as described by the
repositoryCfg file.

The passed-in instantiated mapper and the
reinstantiated mapper do not compare equal,
becuase they do not implement the __eq__ method
and they are not the same instance.

Butler interprets this as the mapper from the cfg
and the passed-in mapper do not match, and it
raises an exception.

The better fix is to not pass an instantiated
mapper to Butler init, instead pass the class
object and let Butler init the mapper.