Rewrite the tests to use the formatter interface
Formatter.read(component) is the interface that everything should
be using so change the tests to use that interface rather than
the internal interface. This is partly driven by V2 no longer
guaranteeing that the file you need to open is the file
that was given to the formatter constructor.